api_news/requirements.txt
Django==4.2.20
djangorestframework==3.15.2
requests
django-cors-headers
gunicorn